+++ src/main/java/com/taller_mecanico/repository/ServicioRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ServicioRepository extends JpaRepository<Servicio, Integer> {
+public interface ServicioRepository extends JpaRepository<Servicio, Long> {
     //Solo servicios activos
     List<Servicio> findByActivoTrue();
